Updates
- Renamed sample
- Added README
- Refactor AppDefaults -> ServiceDefaults
