+++ samples/ClientAppsIntegration/ClientAppsIntegration.AppHost/Properties/launchSettings.json
