+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Components/Layout/MainLayout.razor
