+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Components/Layout/NavMenu.razor.css
