+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Components/Pages/Counter.razor
