+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Components/Pages/Error.razor
