+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Components/Pages/Weather.razor
