+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Components/_Imports.razor
@@ -7,5 +7,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.OutputCaching
 @using Microsoft.JSInterop
-@using ClientAppsServiceDiscovery.Web
-@using ClientAppsServiceDiscovery.Web.Components
+@using ClientAppsIntegration.Web
+@using ClientAppsIntegration.Web.Components
