+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/Properties/launchSettings.json
