+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/appsettings.json
