+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/wwwroot/app.css
