+++ samples/ClientAppsIntegration/ClientAppsIntegration.Web/wwwroot/favicon.png
