+++ samples/ClientAppsIntegration/ClientAppsIntegration.WinForms/MainForm.resx
