+++ samples/ClientAppsIntegration/ClientAppsIntegration.sln
@@ -2,17 +2,22 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34414.90
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsServiceDiscovery.AppHost", "ClientAppsServiceDiscovery.AppHost\ClientAppsServiceDiscovery.AppHost.csproj", "{AF56C3D7-5215-46F6-B9A5-7DC8627A1AAB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsIntegration.AppHost", "ClientAppsIntegration.AppHost\ClientAppsIntegration.AppHost.csproj", "{AF56C3D7-5215-46F6-B9A5-7DC8627A1AAB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsServiceDiscovery.ServiceDefaults", "ClientAppsServiceDiscovery.ServiceDefaults\ClientAppsServiceDiscovery.ServiceDefaults.csproj", "{88B59883-83F3-4215-96BC-24EDC497D84C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsIntegration.ServiceDefaults", "ClientAppsIntegration.ServiceDefaults\ClientAppsIntegration.ServiceDefaults.csproj", "{88B59883-83F3-4215-96BC-24EDC497D84C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsServiceDiscovery.ApiService", "ClientAppsServiceDiscovery.ApiService\ClientAppsServiceDiscovery.ApiService.csproj", "{35286E8E-7BE7-4CAE-B11B-DC9BD23AEABC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsIntegration.ApiService", "ClientAppsIntegration.ApiService\ClientAppsIntegration.ApiService.csproj", "{35286E8E-7BE7-4CAE-B11B-DC9BD23AEABC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsServiceDiscovery.Web", "ClientAppsServiceDiscovery.Web\ClientAppsServiceDiscovery.Web.csproj", "{6F014797-4CC4-4189-947F-5186BBAA3129}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsIntegration.Web", "ClientAppsIntegration.Web\ClientAppsIntegration.Web.csproj", "{6F014797-4CC4-4189-947F-5186BBAA3129}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsServiceDiscovery.WinForms", "ClientAppsServiceDiscovery.WinForms\ClientAppsServiceDiscovery.WinForms.csproj", "{A566A703-4045-4AE5-89D9-643AC7D66BDF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsIntegration.WinForms", "ClientAppsIntegration.WinForms\ClientAppsIntegration.WinForms.csproj", "{A566A703-4045-4AE5-89D9-643AC7D66BDF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClientAppsServiceDiscovery.ClientDefaults", "ClientAppsServiceDiscovery.ClientDefaults\ClientAppsServiceDiscovery.ClientDefaults.csproj", "{594AF12F-28A6-4F0B-8752-A3250729BDA2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientAppsIntegration.AppDefaults", "ClientAppsIntegration.AppDefaults\ClientAppsIntegration.AppDefaults.csproj", "{3997CEA7-914F-4A3E-8955-766B846E72BB}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{866D6E10-1D89-468B-9790-3F333C2307A6}"
+	ProjectSection(SolutionItems) = preProject
+		README.md = README.md
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,10 +45,10 @@ Global
 		{A566A703-4045-4AE5-89D9-643AC7D66BDF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A566A703-4045-4AE5-89D9-643AC7D66BDF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A566A703-4045-4AE5-89D9-643AC7D66BDF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{594AF12F-28A6-4F0B-8752-A3250729BDA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{594AF12F-28A6-4F0B-8752-A3250729BDA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{594AF12F-28A6-4F0B-8752-A3250729BDA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{594AF12F-28A6-4F0B-8752-A3250729BDA2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3997CEA7-914F-4A3E-8955-766B846E72BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3997CEA7-914F-4A3E-8955-766B846E72BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3997CEA7-914F-4A3E-8955-766B846E72BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3997CEA7-914F-4A3E-8955-766B846E72BB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
