--- samples/ClientAppsServiceDiscovery/ClientAppsServiceDiscovery.AppHost/ClientAppsServiceDiscovery.AppHost.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <IsAspireHost>true</IsAspireHost>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ClientAppsServiceDiscovery.ApiService\ClientAppsServiceDiscovery.ApiService.csproj" />
-    <ProjectReference Include="..\ClientAppsServiceDiscovery.Web\ClientAppsServiceDiscovery.Web.csproj" />
-    <ProjectReference Include="..\ClientAppsServiceDiscovery.WinForms\ClientAppsServiceDiscovery.WinForms.csproj"
-                      SkipGetTargetFrameworkProperties="true"
-                      ReferenceOutputAssembly="false" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Aspire.Hosting" />
-  </ItemGroup>
-
-</Project>
